configure postgresql

+++ src/main/java/org/pma/repository/PostRepository.java
@@ -5,16 +5,12 @@ import org.pma.domain.Post;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
-import java.math.BigInteger;
 import java.util.List;
 
 @Repository
-public interface PostRepository extends CrudRepository<Post, BigInteger> {
-
+public interface PostRepository extends CrudRepository<Post, Long> {
 	Post findFirstByOrderByPostedOnDesc();
-
 	List<Post> findAllByOrderByPostedOnDesc();
-
 	Post findBySlug(String slug);
 
 }
